Impliment custom base test case.
